# simpan-uang-api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Simpan Uang Restful API",
    "version": "1",
    "description": "Restful API Documentation for Simpan Uang"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api"
    }
  ],
  "paths": {
    "/user": {
      "get": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Get User",
        "description": "Get user detail",
        "responses": {
          "200": {
            "description": "Success get user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "is_admin": {
                          "type": "boolean"
                        },
                        "balance": {
                          "type": "integer",
                          "format": "int32"
                        },
                        "avatar": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Register User",
        "description": "User registration, first user created is admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "password_confirmation": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success user registration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessMessage"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "User Login",
        "description": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/forgot-password": {
      "post": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "User Forgot Password (only user admin)",
        "description": "Sending generate new password to user email by user admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Reset Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessMessage"
                }
              }
            }
          }
        }
      }
    },
    "/user/change-password": {
      "put": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "User Change Password",
        "description": "User change password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "password_confirmation": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success change password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessMessage"
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Update User Name or Email",
        "description": "Update user name or email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success update user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessMessage"
                }
              }
            }
          }
        }
      }
    },
    "/user/update-avatar": {
      "put": {
        "security": [
          {
            "SimpanUangAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Update Avatar Image User",
        "description": "Update avatar image user using base64",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update Avatar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SimpanUangAuth": {
        "description": "Auth for Simpan Uang Restful API",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "SucessMessage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}